+++ Buzz/QR/QRCodeEditorView.swift
@@ -1,162 +1,6 @@
 import SwiftUI
 import QRCode
 
-// MARK: - Observable Model
-@Observable
-class QRMenuModel {
-  
-  var qrPreviewCornerRadius: CGFloat = 20
-  
-  var urlText: String = "https://example.com" {
-    didSet { generateQRCode() }
-  }
-  
-  var qrForegroundColor: Color {
-    qrBackgroundColor.accessibleTextColor
-  }
-  
-  var qrBackgroundColor: Color = .cyan {
-    didSet { generateQRCode()  }
-  }
-  
-  var viewBackgroundColor: Color {
-    qrBackgroundColor.opacity(0.2)
-  }
-  
-  var selectedPixelIndex: Int = 0 {
-    didSet { generateQRCode() }
-  }
-  
-  var selectedEyeIndex: Int = 0 {
-    didSet { generateQRCode() }
-  }
-  
-  var selectedPupilIndex: Int = 0 {
-    didSet { generateQRCode() }
-  }
-  
-  var qrImage: Image? = nil
-  
-  func generateQRCode() {
-    let fgCGColor = qrForegroundColor.cgColor ?? CGColor(red: 0, green: 0, blue: 0, alpha: 1)
-    let bgCGColor = qrBackgroundColor.cgColor ?? CGColor(red: 1, green: 1, blue: 1, alpha: 1)
-    
-    do {
-      let pixelShape = PixelShapeData.allCases[selectedPixelIndex].makeShape()
-      let eyeShape = EyeShapeData.allCases[selectedEyeIndex].makeShape()
-      let pupilShape = PupilShapeData.allCases[selectedPupilIndex].makeShape()
-      
-      let pngData = try QRCode.build
-        .text(urlText)
-        .quietZonePixelCount(4)
-        .foregroundColor(fgCGColor)
-        .backgroundColor(bgCGColor)
-        .onPixels.shape(pixelShape)
-        .eye.shape(eyeShape)
-        .pupil.shape(pupilShape)
-        .generate
-        .image(dimension: 400, representation: .png())
-      if let uiImg = UIImage(data: pngData) {
-        qrImage = Image(uiImage: uiImg)
-      }
-    } catch {
-      print("Error generating QR code:", error)
-      qrImage = nil
-    }
-  }
-}
-
-// MARK: - Uniform Corner Radius Control
-struct CornerRadiusControl: View {
-  @Binding var cornerRadius: CGFloat
-  var textColor: Color
-  
-  var body: some View {
-    ZStack {
-      // Background Rectangle (Matching Button Style)
-      RoundedRectangle(cornerRadius: 12)
-        .fill(Color.white.opacity(0.2))
-        .frame(height: 50) // Ensuring same height as other controls
-      
-      // Full-Width Slider
-      GeometryReader { geometry in
-        ZStack(alignment: .leading) {
-          // Filled Track
-          RoundedRectangle(cornerRadius: 8)
-            .fill(textColor.opacity(0.5))
-            .frame(
-              width: CGFloat(cornerRadius / 50) * geometry.size.width,
-              height: geometry.size.height - 8
-            )
-            .padding(.horizontal, 4)
-          
-          // Static Label
-          Text("Corner Radius")
-            .font(.headline)
-            .foregroundColor(textColor)
-            .padding(.leading, 12)
-            .frame(height: 50, alignment: .leading)
-        }
-        .gesture(
-          DragGesture()
-            .onChanged { value in
-              let progress = min(max(0, value.location.x / geometry.size.width), 1)
-              cornerRadius = progress * 50
-            }
-        )
-      }
-    }
-    .frame(height: 50) // Standard height
-    .frame(maxWidth: .infinity) // Match button width
-    .padding(.horizontal)
-  }
-}
-
-// MARK: - Corner Radius Preset Sheet
-struct CornerRadiusPresetSheet: View {
-  @Environment(\.dismiss) private var dismiss
-  @State var model: QRMenuModel
-  
-  // Define presets
-  private let presets: [(name: String, value: CGFloat)] = [
-    ("Subtle", 8),
-    ("Medium", 20),
-    ("Full Round", 40)
-  ]
-  
-  var body: some View {
-    NavigationView {
-      HStack(spacing: 16) {
-        ForEach(presets, id: \.name) { preset in
-          Button(action: {
-            model.qrPreviewCornerRadius = preset.value
-            dismiss()
-          }) {
-            RoundedRectangle(cornerRadius: preset.value)
-              .fill(model.qrBackgroundColor)
-              .frame(width: 100, height: 100)
-              .overlay(
-                RoundedRectangle(cornerRadius: preset.value)
-                  .stroke(model.qrForegroundColor, lineWidth: 2)
-              )
-          }
-        }
-      }
-      .padding()
-      .navigationTitle("Corner Radius")
-      .navigationBarTitleDisplayMode(.inline)
-      .toolbar {
-        ToolbarItem(placement: .topBarTrailing) {
-          Button("Done") {
-            dismiss()
-          }
-        }
-      }
-    }
-  }
-}
-
-
 // MARK: - Main QR Code Customizer View
 struct QRCodeEditorView: View {
   @State private var model = QRMenuModel()
@@ -206,11 +50,20 @@ struct QRCodeEditorView: View {
             .frame(width: 250, height: 250)
         }
         
+        
+        
         Spacer()
         
         // Controls
         ScrollView {
           VStack(spacing: 12) {
+            ControlTextField(
+              placeholder: "URL",
+              textColor: buttonTextColor,
+              backgroundColor: model.qrBackgroundColor,
+              text: $model.urlText
+            )
+            
             InlineCustomColorPickerButton(
               title: "Background Color",
               selectedColor: $model.qrBackgroundColor,
@@ -292,7 +145,25 @@ struct QRCodeEditorView: View {
         
         Spacer()
       }
+      // Floating Shuffle Button
+            VStack {
+              Spacer()
+              HStack {
+                Spacer()
+                Button(action: shuffleQRCode) {
+                  Image(systemName: "shuffle")
+                    .font(.system(size: 24))
+                    .foregroundColor(buttonTextColor)
+                    .padding()
+                    .background(model.qrBackgroundColor)
+                    .clipShape(Circle())
+                    .shadow(radius: 5)
+                }
+                .padding()
+              }
+            }
     }
+    .navigationBarTitleDisplayMode(.inline)
     .onAppear {
       model.generateQRCode()
     }
@@ -309,51 +180,63 @@ struct QRCodeEditorView: View {
         .presentationDetents([.medium, .fraction(0.4)])
     }
     .sheet(isPresented: $isShowingCornerRadiusSheet) {
-      CornerRadiusPresetSheet(model: model)
+      CornerRadiusPickerSheet(model: model)
         .presentationDetents([.height(250)])
     }
   }
+  
+  // MARK: - Shuffle Functionality
+    private func shuffleQRCode() {
+      model.qrBackgroundColor = Color(
+        red: Double.random(in: 0...1),
+        green: Double.random(in: 0...1),
+        blue: Double.random(in: 0...1)
+      )
+      model.selectedPixelIndex = Int.random(in: 0..<PixelShapeData.allCases.count)
+      model.selectedEyeIndex = Int.random(in: 0..<EyeShapeData.allCases.count)
+      model.selectedPupilIndex = Int.random(in: 0..<PupilShapeData.allCases.count)
+      model.qrPreviewCornerRadius = CGFloat.random(in: 0...50)
+      model.generateQRCode()
+    }
 }
 
-// MARK: - Inline Custom Color Picker Button
-struct InlineCustomColorPickerButton: View {
-  var title: String
-  @Binding var selectedColor: Color
-  var textColor: Color
-  var backgroundColor: Color // Add this property
+struct ControlTextField: View {
+  let placeholder: String
+  let textColor: Color
+  let backgroundColor: Color
+  @Binding var text: String
   
   var body: some View {
     HStack {
-      Image(systemName: "paintbrush.fill") // Add an icon to match other controls
+      // Leading icon
+      Image(systemName: "link")
         .font(.headline)
         .foregroundColor(textColor)
       
-      Text(title)
-        .font(.headline)
-        .foregroundColor(textColor)
+      // TextField
+      TextField(placeholder, text: $text)
+      .font(.headline)
+      .foregroundColor(textColor)
       
-      Spacer()
-      
-      ZStack {
-        Circle()
-          .fill(selectedColor)
-          .frame(width: 30, height: 30)
-          .overlay(
-            Circle()
-              .stroke(textColor.opacity(0.5), lineWidth: 1)
-          )
-        ColorPicker("", selection: $selectedColor)
-          .labelsHidden()
-          .opacity(0.02)
-          .frame(width: 30, height: 30)
-          .allowsHitTesting(true)
+      // Trailing icon/clear button
+      if !text.isEmpty {
+        Button(action: { text = "" }) {
+          Image(systemName: "xmark.circle.fill")
+            .foregroundColor(textColor)
+        }
+        .frame(width: 30, height: 30)
+        .background(
+          RoundedRectangle(cornerRadius: 8)
+            .fill(backgroundColor)
+        )
       }
     }
     .padding()
     .background(
       RoundedRectangle(cornerRadius: 12)
-        .fill(backgroundColor.opacity(0.5)) // Match ControlButton style
+        .fill(backgroundColor.opacity(0.5))
     )
+    
   }
 }
