+++ Buzz/QR/Shape/PixelShapeData.swift
